src/test/java/ru/yandex/practicum/filmorate/storage/db/MpaDbStorageTest.java
package ru.yandex.practicum.filmorate.storage.db;

import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
import org.springframework.jdbc.core.JdbcTemplate;
import ru.yandex.practicum.filmorate.model.Mpa;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

@JdbcTest
@RequiredArgsConstructor(onConstructor_ = @Autowired)
public class MpaDbStorageTest {
    private final JdbcTemplate jdbcTemplate;

    @Test
    public void testFindMpaById() {
        MpaDbStorage mpaDbStorage = new MpaDbStorage(jdbcTemplate);

        Mpa mpa = mpaDbStorage.getData(1L);

        assertThat(mpa).isNotNull();
    }
}
